requirements.txt
asgiref==3.4.1
Django==3.2.6
django-jazzmin==2.4.8
django-taggit==1.5.1
django-tof==0.3.2
mysqlclient==2.0.3
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
